+++ src/main/java/frc/robot/subsystems/Limelight.java
@@ -4,7 +4,6 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import edu.wpi.first.cameraserver.CameraServer;
 // import edu.wpi.first.math.filter.LinearFilter;
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableEntry;
@@ -26,7 +25,7 @@ public class Limelight extends SubsystemBase {
     public static void pipelineInit() {
         if (team == Alliance.Red)
             pipeline.setDouble(0);
-        else if (team == Alliance.Blue) 
+        else if (team == Alliance.Blue)
             pipeline.setDouble(1);
     }
     public static double getHorizontalOffset() {
@@ -52,13 +51,12 @@ public class Limelight extends SubsystemBase {
     public static double getDistance() {
         double cameraHeight = 0.305; //Height of camera (meters)
         double targetHeight = 0.1455313; //Height of target (meters) measured perfectly
-        double mountAngle = 0; //Angle that the limelight is mounted
+        int mountAngle = 0; //Angle that the limelight is mounted
         return (targetHeight - cameraHeight) / Math.tan(mountAngle + getVerticalOffset()); 
-    } 
+    }
 
     @Override
     public void periodic() {
-        // Values needed from final robot before implemented
         SmartDashboard.putBoolean("Limelight hasValidTarget", hasTarget());
         SmartDashboard.putNumber("Limelight horrizontalOffset", getHorizontalOffset());
         SmartDashboard.putNumber("Limelight verticalOffset", getVerticalOffset());
